used new Assert macro for error handling, added support for mods vec in Amethyst ctx, moved mods class to api and combined it with ModInfo class
