+++ AmethystAPI/src/amethyst/runtime/mod/Mod.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "Json.hpp"
 #include <Windows.h>
 #include <amethyst/Log.h>
 #include <amethyst/Utility.h>
@@ -6,6 +7,7 @@
 #include <iostream>
 #include <shlobj.h>
 #include <vector>
+using json = nlohmann::json;
 
 namespace fs = std::filesystem;
 
@@ -14,12 +16,24 @@ private:
     HMODULE hModule;
 
 public:
+    struct MetaData {
+        std::string name;
+        std::string version;
+        std::vector<std::string> author;
+    };
+
+public:
     std::string modName;
+    MetaData metaData;
 
 public:
     Mod(std::string modName);
     FARPROC GetFunction(const char* functionName);
     void Shutdown();
+
+public:
+    static Mod::MetaData GetMetaData(std::string modName);
+    static Mod::MetaData ParseMetaData(std::string modName, std::string fileContents);
 
 private:
     fs::path GetTempDll();
