--- AmethystAPI/src/amethyst/runtime/mod_info/ModInfo.h
@@ -1,22 +1,0 @@
-#pragma once
-#include "Json.hpp"
-#include "amethyst/Log.h"
-#include <optional>
-#include <string>
-using json = nlohmann::json;
-
-namespace Amethyst {
-    class ModInfo {
-
-    public:
-        struct Meta {
-            std::string name;
-            std::string version;
-            std::vector<std::string> author;
-        };
-
-    public:
-        static std::optional<ModInfo> getModInfoFromText(std::string modName, std::string text);
-        Meta meta;
-    };
-}
