+++ AmethystRuntime/src/AmethystRuntime.cpp
@@ -48,44 +48,12 @@ void AmethystRuntime::ReadLauncherConfig()
 
 void AmethystRuntime::LoadModDlls()
 {
-    std::vector<Amethyst::ModInfo> modInfo = {};
-
     // Load all mods from the launcher_config.json
     for (auto& modName : mLauncherConfig.mods) {
-
-        // Ensure the mod exists
-        std::string modConfigPath = GetAmethystFolder() + "mods/" + modName + "/" + "mod.json";
-
-        if (!fs::exists(modConfigPath)) {
-            Log::Error("[AmethystRuntime] mod.json could not be found, for {}!", modName);
-            throw std::exception();
-        }
-
-        // Try to read it to a std::string
-        std::ifstream modConfigFile(modConfigPath);
-        if (!modConfigFile.is_open()) {
-            Log::Error("[AmethystRuntime] Failed to open mod.json, for {}!", modName);
-            throw std::exception();
-        }
-
-        // Read into a std::string
-        std::stringstream buffer;
-        buffer << modConfigFile.rdbuf();
-        modConfigFile.close();
-        std::string fileContents = buffer.str();
-
-        std::optional<Amethyst::ModInfo> modInfoResult = Amethyst::ModInfo::getModInfoFromText(modName, fileContents);
-
-        // Look if mod info could be read correctly
-        if (modInfoResult.has_value())
-        {
-            modInfo.push_back(modInfoResult.value());
-        }
-
         mLoadedMods.emplace_back(modName);
     }
 
-    this->mAmethystContext.mModInfo = modInfo;
+    this->mAmethystContext.mMods = &mLoadedMods;
 
     // Load all mod functions
     for (auto& mod : mLoadedMods) {
